+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Implemented `malai folder`. You can now share a folder with people without
   having to manually run another HTTP server. Requires `--public` flag as no
   ACL yet, also readonly mode for now.
+- Implemented `malai tcp` and `malai tcp-bridge` to expose any TCP service over
+  kulfi network.
 
 ### malai 0.2.2